chore: Make tests more explicit
