--- tests/dummy/app/adapters/application.js
@@ -1,3 +1,0 @@
-import JSONAPIAdapter from 'ember-data/adapters/json-api';
-
-export default JSONAPIAdapter.extend({});
