+++ tests/dummy/app/serializers/application.js
@@ -1,4 +1,3 @@
 import JSONAPIBagaaravelSerializer from '@bagaar/ember-data-bagaaravel/serializers/json-api-bagaaravel';
-import PaginationSupportSerializerMixin from '@bagaar/ember-data-bagaaravel/mixins/pagination-support-serializer-mixin';
 
-export default JSONAPIBagaaravelSerializer.extend(PaginationSupportSerializerMixin);
+export default JSONAPIBagaaravelSerializer.extend({});
