+++ tests/unit/serializers/json-api-bagaaravel-test.js
@@ -44,40 +44,4 @@ module('Unit | Serializer | json api bagaaravel', function (hooks) {
 
     assert.notOk(existingUserSerialized.data.relationships);
   });
-
-  test('normalizeQueryResponse', async function (assert) {
-    let store = this.owner.lookup('service:store');
-    let Adapter = this.owner.resolveRegistration('adapter:application');
-
-    Adapter.reopen({
-      query() {
-        return Promise.resolve({
-          data: [],
-          links: {
-            first: 'https://api.com/users?page=1&per_page=15',
-            last: 'https://api.com/users?page=9&per_page=15',
-            next: 'https://api.com/users?page=2&per_page=15',
-            previous: null,
-            self: 'https://api.com/users?page=1&per_page=15',
-          },
-          meta: {
-            perPage: 15,
-            total: 125,
-          },
-        });
-      },
-    });
-
-    let users = await store.query('user', {});
-
-    assert.deepEqual(users.meta.pagination, {
-      currentPage: 1,
-      firstPage: 1,
-      lastPage: 9,
-      nextPage: 2,
-      previousPage: null,
-      itemsPerPage: 15,
-      totalItems: 125,
-    });
-  });
 });
